+++ app/src/main/res/layout/fragment_2.xml
@@ -8,7 +8,7 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="Binary"
+        android:text="Binarasdfy"
         android:gravity="center"
         android:textSize="30dp"
         android:textStyle="bold"
